add authentication

+++ src/modules/courses/services/DeleteCourseService.ts
@@ -1,9 +1,11 @@
 import { injectable, inject } from 'tsyringe';
 
+import AppError from '@shared/errors/AppError';
 import ICoursesRepository from '../repositories/ICoursesRepository';
 
 interface IRequest {
   course_id: string;
+  user_id: string;
 }
 
 @injectable()
@@ -13,7 +15,17 @@ export default class ListCoursesService {
     private coursesRepository: ICoursesRepository,
   ) {}
 
-  public async execute({ course_id }: IRequest): Promise<void> {
+  public async execute({ course_id, user_id }: IRequest): Promise<void> {
+    const course = await this.coursesRepository.findById(course_id);
+
+    if (!course) {
+      throw new AppError('Course not found');
+    }
+
+    if (course.author_id !== user_id) {
+      throw new AppError('You can not delete this course');
+    }
+
     await this.coursesRepository.remove(course_id);
   }
 }
